release v3.0x8.7
Sync to isp3x develop branch:

4746611 RkAiqCore: Fix wrong type cast of mTranslator of rk356x
0d9dd19 RkAiqManager: Fix hw_evt_cb not called in single cam mode
8da3a08 aiq: hwi: Sync kernel headers with kernel
1fd19c6 release v3.0x8.7
05d8a39 Revert "aiq: hwi: Sync kernel headers with kernel"
c667b32 aiq: Fix shift bit not enought on 32bit platform
07b5cf1 aiq: hwi: Sync kernel headers with kernel
6e1a093 uapi: remove the uapi enable check for the tool that configure the uapi in real time
5af4b44 uapi: support get 3Astats for each camera.
4f638c9 AF: fix af lock is failed in android
5a40375 alsc: support set/get attr from tool.
39a7419 aie&acp: support set/get attr from tool.
5d26a0d uapi: fix gamma setting.
adea6b9 Fix some bugs in set/get stManual in RK Tool
a799cdd bayertnrV2: delete the enable bit that not used
259fd8d modify nr sharp uapi strength slope factor
a304a53 RPC:fix nr uapi.
8a93245 gic: fix head file missing.
e9294b2 aig: gic: Add j2s comment for uapi struct
fc11d99 Adehaze: support tool in stManual
5ae5d5a csm: add j2s support for set/get attribute
97a6e43 3dlut: add j2s support for set/get attribute
845b282 ccm: add j2s support for set/get attribute
ca69a19 uapi: enable uapi always when using group cam.
2a8c5fc Revert "AE: modify delay scheme"
491e979 fix compile error on Android
1e77cda ipc_server: fix crash when deinit.
4d005e0 iqfiles: update os04a10_DH3588AVS6_default.json
d62bdde AF: add missing af_uapi_head.h
413a6b1 AF: add getCustomAfRes/setCustomAfRes api
9f908f8 AF: fix can not lock af when pdaf is enabled
92ce0ec AF: add spotlt_scene_idx to select meas in spotlight scene
fe635ca AF: improve motor move in each search round
5ba960a AF: change max grid of search position
a4cf358 LENS: add memory clear for set_zoom
46f23eb change the permission of some files from 755 to 644
73ab521 rpc uapi: add support for camgroup.
9d871f1 CaptureRawData: fix wrong precision of snprintf
c4212ed Add interface to get camgroup ctx from rk_aiq_sys_ctx_t
4005c8e reduce error log for RKAIQ running on serial port using 11520 baudrate
8f27b52 nr, sharp, blc & gain modules support manual attri uapi head file
......

+++ xcore/xcam_log.cpp
@@ -171,7 +171,7 @@ void xcam_get_runtime_log_level() {
             g_cam_engine_log_level = strtoull(level, nullptr, 16);
             unsigned long long module_mask = g_cam_engine_log_level >> 12;
             for (int i = 0; i < XCORE_LOG_MODULE_MAX; i++) {
-                if (module_mask & (1 << i)) {
+                if (module_mask & (1ULL << i)) {
                     g_xcore_log_infos[i].log_level = g_cam_engine_log_level & 0xf;
                     g_xcore_log_infos[i].sub_modules = (g_cam_engine_log_level >> 4) & 0xff;
                 }
@@ -196,7 +196,7 @@ int xcam_get_log_level() {
     unsigned long long module_mask = g_cam_engine_log_level >> 12;
 
     for (int i = 0; i < XCORE_LOG_MODULE_MAX; i++) {
-        if (module_mask & ((long long )1 << i)) {
+        if (module_mask & (1ULL << i)) {
             g_xcore_log_infos[i].log_level = g_cam_engine_log_level & 0xf;
             g_xcore_log_infos[i].sub_modules = (g_cam_engine_log_level >> 4) & 0xff;
         }
